limited angular gradient
