+++ crates/grida-canvas/src/shape/ellipse_ring_sector.rs
@@ -1,4 +1,3 @@
-use super::*;
 use math2::bezier_a2c;
 use skia_safe::{Path, Rect};
